Batch event listener calls and rename plugin hook
