+++ src/hooks/useEditorEventListener.ts
@@ -1,9 +1,10 @@
+import { Plugin } from "prosemirror-state";
 import type { DOMEventMap, EditorView } from "prosemirror-view";
 import { useCallback, useContext, useRef } from "react";
 
 import { EditorContext } from "../contexts/EditorContext.js";
-import type { EventHandler } from "../plugins/createReactEventPlugin.js";
 
+import type { EventHandler } from "./useComponentEventListenersPlugin";
 import { useEditorEffect } from "./useEditorEffect.js";
 
 /**
@@ -24,12 +25,14 @@ export function useEditorEventListener<EventType extends keyof DOMEventMap>(
     ref.current = handler;
   }, [handler]);
 
-  const eventHandler = useCallback(
-    (view: EditorView, event: DOMEventMap[EventType]) => {
-      return ref.current(view, event);
-    },
-    []
-  );
+  const eventHandler = useCallback(function (
+    this: Plugin,
+    view: EditorView,
+    event: DOMEventMap[EventType]
+  ) {
+    return ref.current.call(this, view, event);
+  },
+  []);
 
   useEditorEffect(() => {
     registerEventListener(eventType, eventHandler);
